{{cookiecutter.app_name}}/requirements.txt
aioconsole==0.3.1
aiofiles==0.6.0
Chameleon==3.9.0
click==7.1.2
fastapi==0.63.0
fastapi-chameleon==0.1.6
h11==0.12.0
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.15.0
starlette==0.13.6
terminaltables==3.1.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
motor
